ToDos can be deleted Access to songs and todos closes: #3 #5

+++ app/src/main/java/com/danielspeixoto/musician/model/DeleteTaskModel.java
@@ -22,11 +22,11 @@ public class DeleteTaskModel implements IDeleteModel<Task> {
     }
 
     @Override
-    public void delete(int id) {
+    public void delete(long id) {
         SQLiteDatabase db = mDBHandler.getWritableDatabase();
         db.delete(Task.TABLE,
                 Task._ID + " = ?",
-                new String[]{Integer.toString(id)}
+                new String[]{Long.toString(id)}
         );
         mDeletePresenter.onDeleted();
     }
